Why is offset printing not as good as copperplate in book printing?
Why is offset printing not as good as copperplate in book printing?

The following is an analysis by a professional China printing company to explain the definition and difference between them:

1. Coated paper
Coated paper, also known as printing coated paper, is made by coating a layer of white paste on base paper and calendering it. The paper has a smooth surface, high whiteness, uniform paper fiber distribution, uniform thickness, low flexibility, good elasticity, strong water resistance and tensile properties, and good ink absorption and reception. Coated paper is mainly used for printing picture albums, covers, postcards, exquisite product samples and color trademarks. The pressure should not be too high when printing on coated paper, and offset printing resin ink and bright ink should be used. To prevent the back from sticking and dirty, you can use anti-dirty agent, powder spraying and other methods.

2. Offset paper
Offset paper is mainly used for offset (offset) printing machines or other printing machines to print higher-level color prints, such as color pictorials, picture albums, posters, color printing trademarks and some advanced books, as well as book covers, illustrations, etc. Offset paper is divided into special grade, No. 1 and No. 2 according to the proportion of paper pulp. There are single-sided and double-sided, and there are two grades of super calender and ordinary calender.
Offset paper has small stretchability, uniform ink absorption, good smoothness, tight and opaque texture, and good whiteness. Strong water resistance. Conjunctival type offset printing inks and better quality lead printing inks should be used. The viscosity of the ink should not be too high, otherwise the phenomenon of powder removal and fuzzing will occur. Also to prevent dirt on the back side, generally use anti-dirt agent, dusting or interleaving paper.

3. Copperplate printing
In the process of copperplate printing, the artwork is photographed and separated into positive films, which are then exposed and corroded into copper printing plates. The copper printing plate is the opposite of the intaglio plate. The inked part of the pattern is raised while the blank part is recessed.
Advantages: high printing pressure, saturated ink, thick text, and plain background.
Disadvantages: The mesh is relatively coarse, and the fine level is not suitable for plate printing, and is more suitable for printing simple pictures and texts. Copperplate printing expresses the depth and level of the dot size, and the text and background color are not added to the screen. In addition to solid printing, the copperplate pressure is heavier, which often leaves traces of ink spillage at the edges of the print pattern.

4. Offset printing
"Offset printing" and "offset printing" mean copperplate sticker printing. This printing method transfers the ink on the rubber surface to the paper surface through a roller-type rubber stamp. Since the glue surface is flat and there are no concave patterns, the patterns and patterns on the printed paper are also flat, without a three-dimensional effect, and poor anti-counterfeiting. Offset printing requires less ink, and the manufacturing cost of the mold is lower than that of gravure.
Coated paper is not made for copperplate printing, and offset paper is not made for offset printing. It's just a different printing method compared to printed matter and crafts. Can only say which paper is more suitable for which printing method.
If you have decided which way to print your books, please do not hesitate to contact us, We are one of the best printers in China, we will provide you with the best customized China printing service.